+++ JOAO_BECKER/primeirobi/Exercicio2/FloriculturaMelhorada.java
@@ -1,4 +1,4 @@
-package primeirobi.Floricultura2;
+package primeirobi.Exercicio2;
 import java.util.Scanner;
 public class FloriculturaMelhorada {
     public static void main(String[] args){
