+++ JOAO_BECKER/primeirobi/Exercicio3/Floricultura3.java
@@ -1,9 +1,9 @@
-package primeirobi.Floricultura4;
+package primeirobi.Exercicio3;
 
 import java.util.Scanner;
 
-public class Floricultura4 {
-    public static void main(String[] args){
+public class Floricultura3 {
+       public static void main(String[] args){
         try (Scanner leitor = new Scanner(System.in)) {
             boolean loja=false;
             float [][] vendasMes = new float [13][32];
@@ -90,5 +90,4 @@ public class Floricultura4 {
             }
         }
     }
-    
 }
