--- JOAO_BECKER/primeirobi/Exercicio5/Aulaseis.java
@@ -1,13 +1,0 @@
-package primeirobi.Exercicio5;
-
-public class Aulaseis {
-    public static void main(String[] args) {
-         Professor professorum = new Professor();
-
-         professorum.nome = "Alex";
-         professorum.materia = "A.I";
-
-         professorum.apresentarse();
-    }
-    
-}
