+++ JOAO_BECKER/primeirobi/Exercicio5/Vendedor.java
@@ -1,0 +1,30 @@
+package primeirobi.Exercicio5;
+
+public class Vendedor {
+    String nome;
+    Integer idade;
+    String loja;
+    String cidade;
+    String bairro;
+    String rua;
+    Double salarioBase;
+    Double [] salarioRecebido = new Double[3];
+
+    public double calcularMedia(){
+        double media=0;
+        for(int i=0;i<salarioRecebido.length;i++){
+            media=media+salarioRecebido[i];
+        }
+        return media/3;
+    }
+
+    public double calcularBonus(){
+        return salarioBase+(salarioBase*0.2);
+
+    }
+
+    public void apresentarse(){
+        System.out.println("O vendedor "+nome+" da loja "+loja+" tem "+idade+" de idade");
+    }
+}
+
